refactor: set argument u64 instead of conversion

+++ packages/rs-drive-abci/src/execution/validation/state_transition/state_transitions/identity_credit_withdrawal/structure/v0/mod.rs
@@ -48,7 +48,7 @@ impl IdentityCreditWithdrawalStateTransitionStructureValidationV0
         }
 
         // validate core_fee is in fibonacci sequence
-        if !is_fibonacci_number(self.core_fee_per_byte()) {
+        if !is_fibonacci_number(self.core_fee_per_byte() as u64) {
             result.add_error(InvalidIdentityCreditWithdrawalTransitionCoreFeeError::new(
                 self.core_fee_per_byte(),
                 MIN_CORE_FEE_PER_BYTE,
